api-service/api/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1 
torchvision==0.9.1
nltk==3.5
numpy==1.20.1
easydict==1.9
scikit_image==0.18.1
pandas==1.2.3
Pillow==8.2.0
python_dateutil==2.8.1
PyYAML==5.4.1
gunicorn==20.1.0